API update to ErrorHandler.java. ErrorHandler.java now implements UncaughtExceptionHandler. Secondly, remove the questionable "is game over" property in ErrorHandler.java and associated setter calls.
The prior implementation of ErrorHandler did not work. It is set up as a System property with a text based awt even key in GameRunner2. It does not catch swing exceptions, and AFAIK, no exceptions at all. But, it will work if ErrorHandler implements UncaughtExceptionHandler and is set with "Thread.setUncaughtExceptionHandler( Errorhandler )"