API update to ErrorHandler.java. ErrorHandler.java now implements UncaughtExceptionHandler. Secondly, remove the questionable "is game over" property in ErrorHandler.java and associated setter calls.
The prior implementation of ErrorHandler did not work. It is set up as a System property with a text based awt even key in GameRunner2. It does not catch swing exceptions, and AFAIK, no exceptions at all. But, it will work if ErrorHandler implements UncaughtExceptionHandler and is set with "Thread.setUncaughtExceptionHandler( Errorhandler )"

Updates to GameRunner2.java:
 - Add an UncaughtExceptionHandler field to GameRunner2 and set it via Thread.setUncaughtExceptionHandler. When it is not set it defaults to an ErrorHandler.java object, appropriate for end-users. Setter method allows for test clients to inject a test version of the error handler.
 - Improve exception handling when game launch wait window fails
 - Update GameRunner2 private method name "checkForUpdates" to "checkForGameEngineUpdates"
 - Avoid NPE in setupLogging, instead log that the loggers property file could not be loaded.