Add unit test libraries, hamcrest-all and mockito-all.
hamcrest-core is included in hamcrest-all so it is being removed. There are additional hamcrest library jars included with hamcrest-all that include the hamcrest Matchers library. These matchers are a pretty key feature to hamcrest and allows for things like "Matchers.stringContainsInOrder("abc", "123")"

Also the new mockito-all library will allow for mockito mocks. This is an alternative to JMock and EasyMock (so those should not be used) that allows for mocks to be set up with pretty little overhead (including annotation support)