ValidateAttachmentsTest.java - Fix to make sure that a class ends with ".class" before loading it.
The test does a recursive directory search of what it "thinks" is the build folder and checks any classes it finds. In some cases files which are not compiled java class files can wind up in those folders. This class tries to load those files by name, does a substring operation of the file name with the index of ".class". For these unexpected files, that index of is -1 and causes an error. The fix here is to just ignore those files. This test needs overall improvement, it does not assert anything, but at least now it will tolerate non java .class files in the search path defined by the test.