AbstractUndoableMovesPanel.java - Add undo all moves button
Add an "undo all" button that appears in the lower right of the moves panel after the first move has been made (when multiple moves are available to undo). Also part of this change, rename the existing undo action listener to be called a listener. It was previously called just an action and not a listener, so adding in a new listener was very odd in terms of consistent naming (hence the rename for consistency).

Also include some logic in case a move cannot be done. In that case suppress the error message to user and move on and undo any other moves that can be undone.