AbstractUndoableMovesPanel.java - Add 'undo all' moves button
The undo all simply starts from the most recent move and does an undo for that move index. This is taking the same action as clicking the corresponding "undo" button. The undo of all moves avoids index out of bound errors when trying to undo raplidly in network games and helps avoid game delays when all moves are undone (in network games each undo can take a second or more because of lag).

Some more details:
- Button appears when there is more than one move made
- Some plumbing updates, renaming Listeners to be more consistent
- Add logic to suppress an error message when a move cannot be undone, allows for undo-all to effectively ignore those moves.