NewGameChooser - Non-blocking clear of parsed map data.
Allows games to start without having to wait for background map parsing to complete, that data is cleared anyways when a game starts. We do this by making the clear operation non-blocking, we now set a flag to notify refresh operations that they should ignore their results when done computing. Secondly we will also actively cancel refresh operations to avoid unnecessary resource usage.

For context: map data is loaded in the background when tripleA is launched or when a user leaves a game. This data is used to pre-compute the game selection window, letting a user choose a game quickly.  To save memory, we remove this preloaded background data while a user is playing a game. By restarting a local game rapidly, you can notice the delay of loading map data (I had 5 seconds with 11 maps on a powerful 6 core CPU with rotating rotating hard disks)