Backwards compatible fix and update of how test map game data is found. Secondly, make a copy paste of the big world map data in "maps" and move it to "test_data", then update the two tests that used the "maps" version rather than a test version.
The code for finding maps can take a lot of branches, depends on if you run from IDE, ant, gradle, or which packaged version of the game. In addition the test and prod search paths for map data are mixed together. In the test case we find we are running from a class file and assume we are running from ant. In that case all we need to do is ascend up enough directories to get out of the packages and then plus a fixed amount. For gradle this is not enough, the compiled classes are nested a bit deeper.

To fix this, instead of having GameRunner2 tell us what the root folder is, and then from there try to find parent folders with a map file in it, we instead ask GameRunner2 to find us a file and give it a list of potential parent folders. The method that looks at both test and prod paths has been deprecated so we can migrate away from it (so eventually the list woudl perhaps become a scalar). Secondly, GameRunner2 now has different behavior in the finding files in the case of finding relative to a class file (as opposed to relative to a jar). Instead of going up enough directories to match the packages plus one, instead we are given a folder and file, and we keep ascending directories until we find that folder containing the file. This works in both the ant and gradle cases.

Lastly, some extract method was done on GameRunner2, so there is some additional refactoring that was done to clean up the methods so that this change would be clear and make sense. So one can expect some extracted methods in the diff.