Remove occasionally failing test 'testSimple' from ThreadPoolTest
The implementation of ThreadPool will be changing to a library implementation, which reduces the value of this test. In addition, the test still does not pass all the time with a library implementation, indicating something is probably wrong with the test itself. Last, practically speaking, the test failure is being systematically ignored, so the test is providing no value (whiling having a significant cost)