undo_move with u key Move key listeners to MapPanel from TripleAFrame.java
TripleAFrame.java constructs the map arrow key listener for map scrolling and adds that listener to itself and the mapPanel. The one added to the MapPanel is the only valid one, so remove the redundant add of this listener to the TripleAFrame. Furthermore, to gain access to the MovePanel undo key listener, have the TripleAFrame construct the MovePanel instead of ActionButtons creating it. Now TripleAFrame can also add the undo move key listener to the map panel. Until now the undo key listener did not always work, the move panel required an interaction before the undo key would work.