Improve Territory Name Placement
Instead of just finding the bounding rectangle and placing the name in
the center, actually try to find the best location within the territory.
This now finds the rectangle that can fit the name, that is the closest
to the vertical center, and has a large width at that location. If one
isn't found then default back to the bounding rectangle center.