Improve Territory Name Placement (#622)
* Improve Territory Name Placement
Instead of just finding the bounding rectangle and placing the name in
the center, actually try to find the best location within the territory.
This now finds the rectangle that can fit the name, that is the closest
to the vertical center, and has a large width at that location. If one
isn't found then default back to the bounding rectangle center.
* Moved method 'getBestTerritoryNameRect' to IDrawable and made private
since its only used there.
* Created helper method for checking if rectangle is in territory. Updated
increments to create 20x20 grid for each territory. Cleaned up
variables.