On the staging screens, replace usage of the word 'game' when a 'map' is intended.
The code refers to a set of rules + XML file as a game. The code and more generally a game is also the same thing as a map, plus a series of user actions advancing the game, ie: a game-in-progress. What has wound up happening is the word "game" is used mostly everywhere, a bit confusingly when actually just a "map" is being referred to.  This change updates teh staging screens to use the words "game" and "map" more consistently.