Remove 'cancel' option from rocket attack select unit dialog
The cancel button stays on the same screen when selecting rocket damage targets, in effect it does nothing. This is a quick change to remove the button. A good next step/enhancement to the game would be to add the cancel back in, and have it so when cancel is clicked a player can pick a new territory to rocket attack. This looks to be a difficult change at the moment since the territory to attack is pretty baked into the unit target selection. Thus for now we remove the button that does nothing.