Better ThreadUtil.sleep, return boolean and interrupt current thread
Instead of swallowing the interrupted exception we will instead return false. This will allow clients to control the behavior as they would like. As per best practices, we also now interrupt the current thread. When catching an InterruptedException, the thing that read if the thread was interrupted, will have reset the interrupted flag. So to handle that it is best practice to set the interrupted flag again.

Further background and discussion:  https://github.com/triplea-game/triplea/issues/1448