Ensure LockUtil fields are properly synchronized
This change fixes potential non-atomic access of the LockUtil fields.
The previous implementation of LockUtil used static fields to track and
enforce lock ordering.  However, it used an instance field to serialize
access to the static fields.

Within the application, there are at least two instances of LockUtil
(one instance for all Tiles and one instance per GameData).  Thus,
different instances of LockUtil synchronize on a different lock
variable, which does not have the desired effect of serializing access
to the static fields.

Because LockUtil was using static fields to track and enforce lock
ordering, it was effectively a singleton.  This change formalizes that
role and ensures there is only ever a single instance of LockUtil
created.