Replace ValueRange with simply a value range description String
The argument is that the settings currently only support a single type
of value range (IntegerValueRange) and that introducing a new interface
to model an abstract value range at this time is overengineering.

Settings that do not support a value range return a Null Object modelled
as an empty string.