Prevent concurrent access to shared LockUtil fields (#1506)
* Ensure LockUtil fields are properly synchronized

This change fixes potential non-atomic access of the LockUtil fields.
The previous implementation of LockUtil used static fields to track and
enforce lock ordering.  However, it used an instance field to serialize
access to the static fields.

Within the application, there are at least two instances of LockUtil
(one instance for all Tiles and one instance per GameData).  Thus,
different instances of LockUtil synchronize on a different lock
variable, which does not have the desired effect of serializing access
to the static fields.

Because LockUtil was using static fields to track and enforce lock
ordering, it was effectively a singleton.  This change formalizes that
role and ensures there is only ever a single instance of LockUtil
created.

* Ensure custom LockUtil error reporting is only visible for testing

- Extract an interface to represent the error reporter.  This was done
  to avoid any output to stderr during testing.  Previously, when
  creating a spy around ErrorReporter, the default implementation of
  reportError() was called, which output a stack trace to stderr.  This
  could cause devs to think something went wrong while running tests.
- Use Mockito instead of hand-crafted mocks.
- Ensure default error reporter is restored when tests are complete.
- Ensure all access to the error reporter is atomic.

* Rename fields per project code standards

* Reduce visibility of LockUtil.WeakLockRef

* Properly initialize ThreadLocal instead of manually initializing on demand

* Not necessary to synchronize access to ThreadLocal

* Extract methods to more conveniently acquire/release locks via LockUtil

* Implement LockUtil singleton as a single-element enum