Show message dialog immediately if already running on the EDT
This change prevents a deadlock that will occur if this method is
invoked from the EDT.  In that case, the method will block waiting for a
latch to be signaled.  However, the code that signaled the latch is
forked in a SwingUtilities.invokeLater() call.  Thus, the method is
awaiting a signal that will never occur because it is blocking the EDT
from processing new events, which includes the forked latch count down.