Add serialization proxy framework for save games
Some notes from previous commits that were squashed into this commit:

* Persist game data memento instead of game data directly

Introduction of a game data memento is necessary because there are
multiple forms of persisted game data.  In some cases, the complete game
data must be persisted.  In other cases, notably creating a shallow
clone for AI or odds calculation simulations, the game data delegates
must not be persisted.  Therefore, we allow the user to create a memento
for game data that contains only the information that should ultimately
be persisted.  Thus, the memento itself is what is serialized rather
than the game data directly.

* Add missing serialVersionUID fields

Externalizable classes require a serialVersionUID field [1].  However,
Eclipse does not warn when it is missing [2].  I only noticed it because
the OpenJDK compiler issues a warning when "-Xlint:all" is enabled.

[1] http://bugs.java.com/bugdatabase/view_bug.do?bug_id=4094702
[2] https://bugs.eclipse.org/bugs/show_bug.cgi?id=101476