Fix out-of-date map check
The out-of-date map check was using the wrong file name when looking for
a map's download properties.  It only considered the property file
"<mapName>.zip.properties", which does not exist for map's downloaded
from GitHub.  When the map download properties cannot be found, it is
assumed the map is up-to-date.

Since migrating the maps to GitHub, ResourceLoader considers the
following map zip file candidates when attempting to load a map:

* <mapName>.zip
* <normalizedMapName>-master.zip
* <normalizedMapName>.zip

The fix for this issue was to modify the out-of-date map check to also
consider all these map zip file names when attempting to locate the
corresponding download properties file (the same name with ".properties"
appended).