Model download length as Long instead of Integer
This change required more modifications than should have been necessary
because of how the UI progress bar range was modeled.  It previously set
the maximum value (an Integer) as the length of the download.  Now that
the length is modeled a Long, we cannot simply map it one-to-one to the
progress bar range.

The solution was to model the progress bar range as the ratio of
downloaded bytes to total bytes in the range [0,100].  Thus, the
progress listener is now responsible for calculating this ratio when the
count of downloaded bytes has been updated.  In order to calculate the
ratio, it needs to know the total bytes, and this is what required so
many modifications.  The MapDownloadProgressListener class was
introduced to encapsulate this information and logic.