Enforce no duplicate keys in YAML files
SnakeYAML (the parser used by the Jackson YAML converter) defaults to
allowing duplicate keys because the YAML 1.1 spec is not clear on this
issue.  However, this option can be overridden to raise an error when a
duplicate key is detected.

Unfortunately, the Jackson YAML converter does not provide a way to pass
this configuration option to the SnakeYAML parser.  Therefore, we had to
add a second phase to the YAML validation task that uses SnakeYAML
directly to parse each file with the no duplicate key option enabled.

According to the SnakeYAML docs for
LoaderOptions#setAllowDuplicateKeys(), the YAML 1.2 spec makes it clear
that no duplicate keys are allowed.  It is expected that a future major
revision of SnakeYAML will change the default behavior to disallow
duplicate keys.