Nest all related test fixtures under a single outer fixture
Using the Enclosed test runner, we can keep all independent, but
related, test fixtures under a single fixture that provides a namespace
for the tests.  This allows us to achieve the goal of having one test
file per Java class.

Unfortunately, using the Enclosed test runner causes each test to be
executed twice under certain conditions.  This is due to the Enclosed
test runner being essentially a JUnit suite.  The test discovery
protocol of certain tools will identify the suite and the test classes
that make up the suite independently.  Thus, those tests nested within
an Enclosed test runner will be executed twice if not compensated for.

The following sections describe the conditions under which Enclosed
tests may be run twice:

1. Eclipse IDE

Running an Enclosed test directly via Run As > JUnit Test will cause
Eclipse to prompt the user if they want to run the outer (Enclosed)
fixture or one of the nested fixtures.  Choosing to run the outer
fixture will run all nested tests as expected: that is, once.

However, choosing to run JUnit tests for any workspace container that is
an ancestor of the Enclosed test will cause the nested tests to be run
twice.  For example, selecting the package containing an Enclosed test
(or any parent package) and using Run As > JUnit Test will results in
the nested tests being run twice.  It is not known if these duplicate
test runs can be prevented in this scenario.

2. Gradle

Running the test task will result in all nested tests being run twice
because the test discovery protocol picks up both the Enclosed suite as
well as the individual tests.  This is more serious as the test counts
in the CI environment can effectively be double of what they actually
are leading to longer build times and false test metrics.

The solution is to specifically exclude the test discovery protocol from
running the Enclosed suites.  To do this, a unique method for
identifying such tests had to be established.  The convention chosen in
this change is to use the suffix "Tests" instead of the more traditional
"Test" for Enclosed suites.  Gradle is then directed to exclude tests
whose class names end in "Tests".  Thus, as lone as new Enclosed suites
follow this naming convention, no future changes need be made to the
Gradle configuration to prevent duplicate test runs.