Write protocol handler to HKCU\Software\Classes instead of HKCR
HKCR is a virtual key.  Writing to it may actually end up writing to
HKLM if the new key does not already exist under HKCU.  This then
requires the installer to run elevated.  To allow the installer to work
correctly when run unelevated (i.e. a single user install), write the
keys directly to HKCU\Software\Classes.