Fix variable naming violation and simplify loop.
Primarily fix a variable naming violation where variable names end with a dollar sign. While here, the do/while loop can be converted to a while loop. But once that is done, it is pretty obvious it can be converted to a for loop, and then again to a foreach loop removing both the object cast and index variables.