Remove the option for delayed parsing.
First what the option does. This option would take effect when loading the list of maps available for selection. When delayed parsing is on, we do a 'shallow' parse of the XML files which is faster than a 'deep' or 'full' parse. When delayed parsing is off, we do a 'full' parsing of every game on the list.

Full parsing can be problematic when 'War of the Relics' is on the list, since that map takes a few minutes to parse. In previous conversations about that map, this feature did not have much value even to map makers. Hence, the removal, this is something that should not be changed from the default.