Lobby start update, init DB first
Currently the lobby start up sequencing is lobby service first, then init and start database. A problem here is that if there is an issue starting or connecting to database, the lobby service is left in a running but non-functional state. Launching the app again then fails due to a port collision. In production contexts this can be problematic if we fail to notice database issues. In this case we could launch the service, see it is running, and potentially walk away.

The update here simply changes the start up order, so we init DB first and then lobby. In this case if there are problems with database we will have a failure to start, and the lobby service will never launch. Not only is it easier to notice and monitor for this problem, and cleanup/recovery is simpler (pertinent in dev context).