Validate map download URLs
It was discovered that AsyncHttpClient converts a HEAD request to a GET
after a redirect [1].  Although this behavior is incorrect per RFC2616,
the author made a conscious decision to implement it this way to match
the behavior of major browsers.

Unfortunately, the switch to a GET causes our URL validation to actually
download the resources instead of simply probing them via a HEAD.  This
wasn't a big deal when we were only validating thumbnail URLs, but now
that we are validating map URLs, the download size is hundreds of MiB!

One can provide a custom handler to AsyncHttpClient that will abort the
transfer after the headers are received.  However, I found doing that
caused both SourceForge and GitHub to reset the connection quite often,
resulting in spurious build failures.

I ultimately chose to go back to using the Apache HTTP client, which
does preserve the HEAD method upon a redirect.  However, the Apache
client does not support multiple asynchronous requests.  We have a
couple of hundred URLs to validate and running the probes in parallel
reduces the duration of this task by approximately a factor of two.
Therefore, I parallelized the probes using a vanilla thread pool.

The only side effect of this change is that the validation no longer
fails fast.  It collects all the probe results before checking them.  If
this becomes an issue, we can revisit changing the implementation in the
future.

[1] https://github.com/AsyncHttpClient/async-http-client/issues/989