Make installer register triplea as handler for triplea: domains (Partly #1723) (#1783)
* Added Windows registering

* Add Linux protocol handler

* Fix java code (was invalid home dir)

* Fix uninstall rights on windows

* Pass + Use params on linux systems

* Upgrade install4j to 6.1.5

* Fix error message casing

* Write protocol handler to HKCU\Software\Classes instead of HKCR

HKCR is a virtual key.  Writing to it may actually end up writing to
HKLM if the new key does not already exist under HKCU.  This then
requires the installer to run elevated.  To allow the installer to work
correctly when run unelevated (i.e. a single user install), write the
keys directly to HKCU\Software\Classes.
* Fix scripts
* Solve os check in an elegant way