Update lobby server Eclipse launch configurations
The lobby no longer supports a UI, so it is not necessary to have two
separate launch configurations for with and without a UI.

I also removed the directive to ignore *.launch files in the Git repo.
Such a directive makes it difficult to track legitimate changes such as
this one.  I believe the intention was to ignore one-off launch
configurations created by the dev in their personal workspace.  However,
later versions of Eclipse do not store such launch configurations in the
project if the user does not explicitly "share" them.  Rather, they are
stored in the workspace's .metadata folder, which is typically outside
the bounds of source control.