Do not protect credentials in save games
When protected credentials are present in a save game, the player that
receives the save game will not be able to unprotect them, resulting in
an error being logged quietly and the receiving user's credentials will
be cleared.  To avoid this issue, we simply do not protect credentials
in a save game so no unprotection is required on the receiving end.

The "Remember my password" option is now "Remember my credentials"
(rendered as "Remember me" in the UI) because we now allow the user to
opt-out of storing both their user name and password.  This also allows
us to prevent storing the actual user name in the save game.  A use case
was described in which the login name and display name on
AxisAndAllies.org can be different.  By storing the user name in the
save game, the other players can discover the poster's login name,
whereas they may have only previously been known by their display name.
Thus, we now protect a user's privacy by not disclosing the user name
outside of the local machine.