Do not persist credentials when using TripleAForumPoster
The `username` and `password` fields were being persisted to the local
cache.  This isn't as big of a deal now that credentials are encrypted,
but it represents another attack surface which we have to worry about in
the future.

These fields were only being used to share state between a few methods.
The fields have been removed, and the information contained therein is
now passed explicitly as an argument to the methods that require them.