Fix race condition in BackgroundTaskRunner#runInBackground()
The root cause of the Download Maps window appearing behind the main
frame is a race condition in the BackgroundTaskRunner#runInBackground()
method.  The race condition is caused by forcing the caller to trigger
completion actions that must be run on the EDT from the background task.
This causes those completion actions to be potentially scheduled before
the completion actions scheduled by runInBackground() itself.

In the Download Maps case, the caller's completion action is to display
the Download Maps window and to bring it to the foreground.  The
runInBackground() completion action is to hide the progress dialog,
which has the side effect of returning focus back to the main frame.
Thus, if the caller's completion action is scheduled to run before the
runInBackground() completion action, the Download Maps window will first
be displayed in the foreground followed immediately by the main frame
being displayed in the foreground giving the appearance that the
Download Maps window was displayed in the background.

The solution is to break out any completion actions the caller needs to
run from the background action.  This then allows runInBackground() to
ensure the progress dialog is hidden (and returning focus to the main
frame) before running the caller's completion action.