Fix temporary progress bar reset when download is complete
The download progress bar will temporarily be reset to 0% after the
download is complete because the FileSizeWatcher is still monitoring the
file size when we attempt to move the intermediate file to its final
location.  If the FileSizeWatcher background thread executes immediately
after the move, it will observe a zero file size, and the progress bar
is reset to 0% for a short time (hundreds of milliseconds) before it
jumps back to 100% when the downloadStopped() callback is invoked.

The fix is to stop the FileSizeWatcher immediately after the download so
its background thread is no longer running when we attempt to move the
file.  Because we don't wait for confirmation that the background thread
has exited before proceeding with the file move, there is a very small
possibility that it still may see a zero file size causing the progress
bar to once again temporarily be set to 0%.