Move required persistence delegate functionality directly into the proxy
The only functionality of a persistence delegate that is currently used
is that of replacing a principal with a serializable proxy.  The other
three features (annotateClass, resolveClass, resolveObject) were unused.
To simplify the code, the replaceObject functionality of the persistence
delegate was moved to the proxy itself in the form of a factory object.

The new architecture is similar to the previous version, but some of the
entity names have changed:

* Proxy --> Proxy (unchanged)
* PersistenceDelegate --> ProxyFactory
* PersistenceDelegateRegistry --> ProxyFactoryRegistry

In addition, because we no longer support the resolveClass and
resolveObject functionality, there is no need for a custom
ObjectInputStream class.  The custom class has been removed and replaced
with the java.io version.