Improve P2P network game authentication protocol
Based on work originally in #2101.

A second (V2) P2P network game authentication protocol was added to
improve security relative to the original (V1) protocol.

The V1 protocol simply hashed the shared game password and a salt
generated by the server to verify that the client knows the game
password. The protocol uses MD5 as the hash function, which is
susceptible  to collision attacks.

The V2 protocol uses a keyed-hash message authentication code (HMAC) of
a secret key (derived from the shared game password and a salt generated
by the server) and a nonce generated by the server to verify that the
client knows the game password. The protocol uses PBKDF2 to derive the
secret key and SHA-512 as the HMAC hash function.

The ClientLogin and ClientLoginValidator classes were heavily refactored
to cleanly support using the V1 and V2 protocols in parallel.