Redesign ProxyFactoryRegistry to ProxyRegistry
With only one responsibility in ProxyFactory, the former
ProxyFactoryRegistry#getProxyFactory() method seemed to lead to a Law of
Demeter violation with calls like:

  proxyFactoryRegistry.getProxyFactory(obj.getClass()).newProxyFor(obj);

It also required us to deal with previously returning
Optional<ProxyFactory> when no proxy factory was available, which has
since been transformed into returning the identity proxy factory
instead.

With this change, that call has been replaced with
ProxyRegistry#newProxyFor().  Now all the logic about whether or not a
proxy factory has been registered is encapsulated within the
ProxyRegistry type, so that the previous example can be replaced with:

  proxyRegistry.getProxyFor(obj);