Fix broken moderator commands when connected to old lobby
The new "zz" methods added in #1942 do not exist in the currently
deployed lobby.  Therefore, any attempt to run a moderator command from
a new client will result in a failure.  The failure is manifest as a
freeze in the UI as the call to the associated IModeratorController
method never returns.