Move proxy registry singleton to a separate class
ClientContext should not have to be modified every time a new
serialization proxy is added to the system.  The new location was chosen
simply because it is co-located with the only current client
(GameDataManager).  As this feature becomes more defined, the most
appropriate location for this singleton will become more apparent.