Add interface for serializable proxies
This interface and its associated documentation is intended to make it
easier for devs to implement their own serializable proxies.

An interface was used instead of an abstract base class to avoid
coupling in the serialization stream.  That is, if a proxy were to
extend a framework class (e.g. AbstractProxy), a descriptor for that
framework class will be written to the serialization stream.  That would
prevent refactoring of the framework class (e.g. moving it to a
different package) because it would otherwise break stream
compatibility.  Class descriptors for interfaces are not written to the
serialization stream, thus making interfaces immune to these issues.